Also refresh PayWay orders for not successful messages

+++ delivery-example/src/main/java/org/socialhistoryservices/delivery/reproduction/service/ReproductionService.java
@@ -49,7 +49,7 @@ public interface ReproductionService {
      * @param id Id of the Order to retrieve.
      * @return The Order matching the id.
      */
-    public Order getOrderById(int id);
+    public Order getOrderById(long id);
 
     /**
      * Retrieve the ReproductionStandardOption matching the given id.
